Resolving NA vs NaN Issues
Many of the uploaded data sets had conflicting entries. Some used NaN while others used NA. Made it consistent such that all use NaN and added in snippet of code to ensure that if a file is loaded in with NA, an error will be thrown to inform the TAs.